requirements.txt
APScheduler==3.7.0
async-lru==1.0.2
pyaes==1.6.1
Pyrogram==1.2.9
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
TgCrypto==1.2.2
tzlocal==2.1
